+++ engine/src/main/java/dev/gamekit/ui/widgets/Parent.java
@@ -14,8 +14,20 @@ public abstract class Parent extends Widget {
   }
 
   /**
-   * Called in the {@link #render} method to provide the parent's appearance before its
-   * child/children are also rendered on top of it.
+   * Called at the start of {@link #render} before {@link #renderAppearance}.
+   * Subclasses can override this to perform any pre-render actions
+   */
+  protected void preRender(Graphics2D g) { /* No-op */ }
+
+  /**
+   * Called after {@link #preRender} method to draw the parent's appearance before its child/children are
+   * also rendered on top of it.
    */
   protected void renderAppearance(Graphics2D g) { /* No-op */ }
+
+  /**
+   * Called at the end of {@link #render} to perform any post-render actions.
+   * Subclasses can override this to perform any pre-render actions
+   */
+  protected void postRender(Graphics2D g) { /* No-op */ }
 }